Create dbupgdiag.sql
This script is intended to provide a user friendly output to diagnose the status of the database either before (or) after upgrade. The script will create a file called db_upg_diag_<sid>_<timestamp>.log.

On Container Database , this should be run individually on each container. 
The script needs to be run in SQL*Plus both before the upgrade on the source database and after the upgrade on the upgraded database as SYS user.

This will help to determine the status of the database before and after upgrade.